+++ src/MatrixComputations.h
@@ -24,7 +24,7 @@ void 	print_matrix(gsl_matrix *M);
 void 	convert_to_gsl(T_FILE_DIM* file_dim, gsl_matrix *M);
 void 	print_vector (gsl_vector *V);
 void    add_submatrix(gsl_matrix* R, gsl_matrix* sub_matrix, uint8 index1, uint8 index2);
-void    delete_column(gsl_matrix* R, uint8 col);
+void    delete_column(gsl_matrix* R, uint16 col);
 double  euclidean_norm(gsl_vector* V);
 boolean compute_matrix_inverse(gsl_matrix * M);
 boolean product_matrix_vector(gsl_matrix* M, gsl_vector* v, gsl_vector* result);
